requirements.txt
# requirements.txt

streamlit
pandas
supabase
python-dotenv
openpyxl